The naming scheme for chunk files is rather unfortunate since colons aren't allowed in filenames on Windows.
My ghetto solution is to add a flag to the config which will turn the colons into dots if set. It's set on automatically on ActiveState Perl on Windows unless explicitly set to off in the configuration file. It can also be explicitly set to on for other systems in case, for some bizarre reason, Brackup is running on a UNIX system but writing its data to an SMB mount from a Windows box.

git-svn-id: http://brackup.googlecode.com/svn/trunk@75 c3152786-d43a-0410-8499-09eb43983aed