Move warning about GTop not being enabled to stats output
Don't spew a warning to STDERR about something which is an optional
dependency, and whose lack in no way negatively affects the
functioning of the program.  Instead, move the warning to the
statistics output, where it is potentially relevant.  Additionally,
mention that GTop would provide memory usage statistics, which is not
necessarily obvious to the casual user.